// dashboard/src/utils/hasura-api/generated/schemas/actionDefinition.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { ActionDefinitionKind } from './actionDefinitionKind';
import type { ActionDefinitionType } from './actionDefinitionType';
import type { Headers } from './headers';
import type { InputArgument } from './inputArgument';
import type { RequestTransformation } from './requestTransformation';
import type { ResponseTransformation } from './responseTransformation';

/**
 * Definition of an action
 */
export interface ActionDefinition {
  /** URL of the action handler. A String value which supports templating environment variables enclosed in {{ and }}. */
  handler: string;
  /** GraphQL type of the action response. */
  output_type: string;
  /** Input arguments of the action */
  arguments?: InputArgument[];
  /** Type of the action */
  type?: ActionDefinitionType;
  /** Execution kind of a mutation action */
  kind?: ActionDefinitionKind;
  /** List of headers to be sent with the handler request */
  headers?: Headers;
  /** If set to true, the client headers are forwarded to the action handler. */
  forward_client_headers?: boolean;
  /** Headers that are not forwarded to the action handler when forward_client_headers is enabled. */
  ignored_client_headers?: string[];
  /** Number of seconds to wait for the handler response before timing out. */
  timeout?: number;
  request_transform?: RequestTransformation;
  response_transform?: ResponseTransformation;
  [key: string]: unknown;
}